+++ src/ragas/metrics/_goal_accuracy.py
@@ -6,8 +6,8 @@ from dataclasses import dataclass, field
 from pydantic import BaseModel, Field
 
 from ragas.dataset_schema import MultiTurnSample
-from ragas.experimental.prompt import PydanticPrompt
 from ragas.metrics.base import MetricType, MetricWithLLM, MultiTurnMetric
+from ragas.prompt import PydanticPrompt
 
 if t.TYPE_CHECKING:
     from langchain_core.callbacks.base import Callbacks
