+++ src/ragas/metrics/_instance_specific_rubrics.py
@@ -4,7 +4,6 @@ import typing as t
 from dataclasses import dataclass, field
 
 from ragas.dataset_schema import MultiTurnSample, SingleTurnSample
-from ragas.experimental.prompt import PydanticPrompt
 from ragas.metrics._domain_specific_rubrics import (
     MultiTurnWithoutReferenceInput,
     MultiTurnWithoutReferencePrompt,
@@ -20,6 +19,7 @@ from ragas.metrics.base import (
     MultiTurnMetric,
     SingleTurnMetric,
 )
+from ragas.prompt import PydanticPrompt
 
 if t.TYPE_CHECKING:
     from langchain_core.callbacks import Callbacks
