+++ src/ragas/metrics/_topic_adherence.py
@@ -8,8 +8,8 @@ import numpy as np
 from pydantic import BaseModel, Field
 
 from ragas.dataset_schema import MultiTurnSample
-from ragas.experimental.prompt import PydanticPrompt
 from ragas.metrics.base import MetricType, MetricWithLLM, MultiTurnMetric
+from ragas.prompt import PydanticPrompt
 
 if t.TYPE_CHECKING:
     from langchain_core.callbacks import Callbacks
