+++ src/ragas/testset/synthesizers/__init__.py
@@ -2,7 +2,10 @@ import typing as t
 
 from ragas.llms import BaseRagasLLM
 
-from .abstract_query import AbstractQuerySynthesizer, ComparativeAbstractQuerySynthesizer
+from .abstract_query import (
+    AbstractQuerySynthesizer,
+    ComparativeAbstractQuerySynthesizer,
+)
 from .base import BaseSynthesizer
 from .specific_query import SpecificQuerySynthesizer
